HIIT-022 refactor: BasicIt 위치 이동 반영
