+++ src/main/java/com/hiit/api/domain/model/it/registered/RegisteredIt.java
@@ -1,5 +1,6 @@
 package com.hiit.api.domain.model.it.registered;
 
+import com.hiit.api.domain.model.it.BasicIt;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
