+++ src/test/java/com/hiit/api/domain/service/it/BrowseItsServiceTest.java
@@ -6,7 +6,7 @@ import static org.mockito.Mockito.when;
 
 import com.hiit.api.AppMain;
 import com.hiit.api.domain.dao.it.registerd.RegisteredItDao;
-import com.hiit.api.domain.model.it.registered.BasicIt;
+import com.hiit.api.domain.model.it.BasicIt;
 import com.hiit.api.domain.usecase.it.RegisteredItEntityConverter;
 import com.hiit.api.domain.util.JsonConverter;
 import com.hiit.api.repository.entity.business.it.RegisteredItEntity;
